moving things around post merge.
