+++ src/Microsoft.ML.Data/Transforms/TransformsExtensionsCatalog.cs
@@ -1,11 +1,13 @@
-﻿using Microsoft.ML.Runtime;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.ML.Runtime;
 using Microsoft.ML.Runtime.Data;
 using Microsoft.ML.Transforms;
 
 namespace Microsoft.ML
 {
-    using HashDefaults = HashEstimator.Defaults;
-
     /// <summary>
     /// Extensions for Column Copying Estimator.
     /// </summary>
@@ -34,7 +36,7 @@ namespace Microsoft.ML
     /// <summary>
     /// Extension ColumnConcatenatingEstimator
     /// </summary>
-    public static class ColumnConcatenatingEstimator
+    public static class ColumnConcatenatingEstimatorCatalog
     {
         /// <summary>
         /// Concatenates two columns together.
@@ -42,52 +44,8 @@ namespace Microsoft.ML
         /// <param name="catalog">The transform's catalog.</param>
         /// <param name="inputColumns">The names of the columns to concatenate together.</param>
         /// <param name="outputColumn">The name of the output column.</param>
-        public static ConcatEstimator Concatenate(this TransformsCatalog catalog, string outputColumn, params string[] inputColumns)
-            => new ConcatEstimator(CatalogUtils.GetEnvironment(catalog), outputColumn, inputColumns);
+        public static ColumnConcatenatingEstimator Concatenate(this TransformsCatalog catalog, string outputColumn, params string[] inputColumns)
+            => new ColumnConcatenatingEstimator(CatalogUtils.GetEnvironment(catalog), outputColumn, inputColumns);
 
-    }
-
-    public static class HashingEstimatorCatalog
-    {
-        /// <summary>
-        /// Hashes the values in the input column.
-        /// </summary>
-        /// <param name="catalog">The transform's catalog.</param>
-        /// <param name="inputColumn">Name of the input column.</param>
-        /// <param name="outputColumn">Name of the column to be transformed. If this is null '<paramref name="inputColumn"/>' will be used.</param>
-        /// <param name="hashBits">Number of bits to hash into. Must be between 1 and 31, inclusive.</param>
-        /// <param name="invertHash">Limit the number of keys used to generate the slot name to this many. 0 means no invert hashing, -1 means no limit.</param>
-        public static HashEstimator Hash(this TransformsCatalog catalog, string inputColumn, string outputColumn = null,
-            int hashBits = HashDefaults.HashBits, int invertHash = HashDefaults.InvertHash)
-            => new HashEstimator(CatalogUtils.GetEnvironment(catalog), inputColumn, outputColumn, hashBits, invertHash);
-
-        /// <summary>
-        /// Hashes the values in the input column.
-        /// </summary>
-        /// <param name="catalog">The transform's catalog.</param>
-        /// <param name="columns">Description of dataset columns and how to process them.</param>
-        public static HashEstimator Hash(this TransformsCatalog catalog, params HashTransformer.ColumnInfo[] columns)
-            => new HashEstimator(CatalogUtils.GetEnvironment(catalog), columns);
-
-    }
-
-    public static class ToValueCatalog
-    {
-        /// <summary>
-        /// Convert the key types back to their original values.
-        /// </summary>
-        /// <param name="catalog">The categorical transform's catalog.</param>
-        /// <param name="inputColumn">Name of the input column.</param>
-        public static KeyToValueEstimator ToValue(this TransformsCatalog.CategoricalTransforms catalog, string inputColumn)
-            => new KeyToValueEstimator(CatalogUtils.GetEnvironment(catalog), inputColumn);
-
-        /// <summary>
-        ///  Convert the key types (name of the column specified in the first item of the tuple) back to their original values
-        ///  (named as specified in the second item of the tuple).
-        /// </summary>
-        /// <param name="catalog">The categorical transform's catalog</param>
-        /// <param name="columns">The pairs of input and output columns.</param>
-        public static KeyToValueEstimator ToValue(this TransformsCatalog.CategoricalTransforms catalog, params (string input, string output)[] columns)
-             => new KeyToValueEstimator(CatalogUtils.GetEnvironment(catalog), columns);
     }
 }
